BE/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-docx
requests
pillow
pytesseract
openpyxl